added the wrapper class for FirebaseAuth and refactored the createUser() method
